requirements.txt
alembic==1.13.1
alembic_utils==0.8.2
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
fastapi==0.110.0
flupy==1.2.0
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httptools==0.6.1
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.5
packaging==24.0
parse==1.20.1
passlib==1.7.4
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic_core==2.16.3
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
